chore(storages): move index/prunner/cache to storages/common
