+++ Cargo.toml
@@ -41,14 +41,15 @@ members = [
     "src/query/pipeline/transforms",
     "src/query/settings",
     "src/query/sql",
-    "src/query/storages/cache",
     "src/query/storages/common/blocks",
+    "src/query/storages/common/cache",
+    "src/query/storages/common/index",
     "src/query/storages/common/table-meta",
+    "src/query/storages/common/pruner",
     "src/query/storages/factory",
     "src/query/storages/fuse",
     "src/query/storages/hive/hive",
     "src/query/storages/hive/hive-meta-store",
-    "src/query/storages/index",
     "src/query/storages/information-schema",
     "src/query/storages/memory",
     "src/query/storages/null",
@@ -58,7 +59,6 @@ members = [
     "src/query/storages/system",
     "src/query/storages/view",
     "src/query/storages/parquet",
-    "src/query/storages/pruner",
     "src/query/users",
     # databend-query
     "src/query/service",
