+++ src/query/service/tests/it/storages/fuse/block_writer.rs
@@ -21,9 +21,9 @@ use common_storages_fuse::io::write_data;
 use common_storages_fuse::io::TableMetaLocationGenerator;
 use common_storages_fuse::io::WriteSettings;
 use common_storages_fuse::FuseStorageFormat;
-use common_storages_index::BlockFilter;
 use opendal::Operator;
 use storages_common_blocks::blocks_to_parquet;
+use storages_common_index::BlockFilter;
 use storages_common_table_meta::meta::BlockMeta;
 use storages_common_table_meta::meta::ClusterStatistics;
 use storages_common_table_meta::meta::Compression;
