+++ src/query/storages/common/cache/src/object.rs
