+++ src/query/storages/common/cache/tests/it/by_pass_cache.rs
@@ -17,18 +17,47 @@ use std::sync::Arc;
 use common_base::base::tokio;
 use common_base::base::uuid;
 use common_exception::Result;
-use common_storages_cache::CacheSettings;
-use common_storages_cache::CachedObject;
-use common_storages_cache::CachedObjectAccessor;
-use common_storages_cache::MemoryItemsCache;
 use opendal::services::fs;
 use opendal::services::fs::Builder;
 use opendal::Operator;
+use storages_common_cache::ByPassCache;
+use storages_common_cache::CacheSettings;
+use storages_common_cache::CachedObject;
+use storages_common_cache::CachedObjectAccessor;
 
-#[derive(serde::Serialize, serde::Deserialize, PartialEq, Eq, Debug, Clone)]
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn test_by_pass_bytes_cache() -> Result<()> {
+    let mut builder: Builder = fs::Builder::default();
+    builder.root("/tmp");
+    let op: Operator = Operator::new(builder.build()?);
+
+    let path = uuid::Uuid::new_v4().to_string();
+    let object = op.object(&path);
+
+    // Cache.
+    let settings = CacheSettings::default();
+    let cache = Arc::new(ByPassCache::create(&settings));
+
+    let expect: Arc<Vec<u8>> = Arc::new("hello, by pass".into());
+
+    // Cached Object Accessor.
+    let accessor = CachedObjectAccessor::create(cache.clone());
+    accessor.write(&object, expect.clone()).await?;
+
+    // Reader.
+    let actual: Arc<Vec<u8>> = accessor.read(&object, 0, expect.len() as u64).await?;
+
+    assert_eq!(actual, expect);
+
+    // Remove.
+    accessor.remove(&object).await?;
+
+    Ok(())
+}
+
+#[derive(serde::Serialize, serde::Deserialize, PartialEq, Eq, Debug)]
 struct TestMeta {
-    a: u64,
-    b: u64,
+    a: u8,
 }
 
 impl CachedObject for TestMeta {
@@ -44,7 +73,7 @@ impl CachedObject for TestMeta {
 }
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn test_memory_items_cache() -> Result<()> {
+async fn test_by_pass_item_cache() -> Result<()> {
     let mut builder: Builder = fs::Builder::default();
     builder.root("/tmp");
     let op: Operator = Operator::new(builder.build()?);
@@ -54,15 +83,16 @@ async fn test_memory_items_cache() -> Result<()> {
 
     // Cache.
     let settings = CacheSettings::default();
-    let cache = Arc::new(MemoryItemsCache::<TestMeta>::create(&settings));
+    let cache = Arc::new(ByPassCache::create(&settings));
 
-    let expect = Arc::from(TestMeta { a: 0, b: 0 });
+    let expect: Arc<TestMeta> = Arc::new(TestMeta { a: 3 });
 
     // Cached Object Accessor.
     let accessor = CachedObjectAccessor::create(cache.clone());
     accessor.write(&object, expect.clone()).await?;
 
-    let actual = accessor.read(&object, 0, 16).await?;
+    // Reader.
+    let actual: Arc<TestMeta> = accessor.read(&object, 0, 8).await?;
 
     assert_eq!(actual, expect);
