+++ src/query/storages/common/index/src/index_min_max.rs
