+++ src/query/storages/common/index/tests/it/filters/bloom_filter.rs
@@ -33,8 +33,8 @@ use common_expression::TableField;
 use common_expression::TableSchema;
 use common_expression::Value;
 use common_functions::scalars::BUILTIN_FUNCTIONS;
-use common_storages_index::BlockFilter;
-use common_storages_index::FilterEvalResult;
+use storages_common_index::BlockFilter;
+use storages_common_index::FilterEvalResult;
 use storages_common_table_meta::meta::BlockFilter as LatestBloom;
 use storages_common_table_meta::meta::Versioned;
