+++ src/query/storages/common/pruner/src/range_pruner.rs
@@ -18,7 +18,7 @@ use common_catalog::table_context::TableContext;
 use common_exception::Result;
 use common_expression::Expr;
 use common_expression::TableSchemaRef;
-use common_storages_index::RangeFilter;
+use storages_common_index::RangeFilter;
 use storages_common_table_meta::meta::StatisticsOfColumns;
 
 pub trait RangePruner {
