+++ src/query/storages/factory/Cargo.toml
@@ -17,10 +17,11 @@ common-config = { path = "../../config" }
 common-exception = { path = "../../../common/exception" }
 common-meta-app = { path = "../../../meta/app" }
 common-storages-fuse = { path = "../fuse" }
-common-storages-index = { path = "../index" }
 common-storages-memory = { path = "../memory" }
 common-storages-null = { path = "../null" }
 common-storages-random = { path = "../random" }
 common-storages-view = { path = "../view" }
+
+storages-common-index = { path = "../common/index" }
 
 dashmap = "5.4"
