+++ src/query/storages/fuse/Cargo.toml
@@ -28,10 +28,10 @@ common-pipeline-transforms = { path = "../../pipeline/transforms" }
 common-sharing = { path = "../../sharing" }
 common-sql = { path = "../../sql" }
 common-storage = { path = "../../../common/storage" }
-common-storages-index = { path = "../index" }
-common-storages-pruner = { path = "../pruner" }
 
 storages-common-blocks = { path = "../common/blocks" }
+storages-common-index = { path = "../common/index" }
+storages-common-pruner = { path = "../common/pruner" }
 storages-common-table-meta = { path = "../common/table-meta" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
