+++ src/query/storages/fuse/src/operations/fuse_sink.rs
@@ -26,9 +26,9 @@ use common_expression::BlockCompactThresholds;
 use common_expression::DataBlock;
 use common_expression::TableSchemaRef;
 use common_pipeline_core::processors::port::OutputPort;
-use common_storages_index::*;
 use opendal::Operator;
 use storages_common_blocks::blocks_to_parquet;
+use storages_common_index::*;
 use storages_common_table_meta::caches::CacheManager;
 use storages_common_table_meta::meta::ColumnId;
 use storages_common_table_meta::meta::ColumnMeta;
