+++ src/query/storages/fuse/src/operations/mutation/compact/compact_transform.rs
@@ -26,9 +26,9 @@ use common_exception::Result;
 use common_expression::BlockCompactThresholds;
 use common_expression::DataBlock;
 use common_expression::TableSchemaRef;
-use common_storages_index::BlockFilter;
 use opendal::Operator;
 use storages_common_blocks::blocks_to_parquet;
+use storages_common_index::BlockFilter;
 use storages_common_table_meta::caches::CacheManager;
 use storages_common_table_meta::meta::BlockMeta;
 use storages_common_table_meta::meta::SegmentInfo;
