+++ src/query/storages/fuse/src/operations/mutation/deletion/deletion_meta.rs
@@ -19,7 +19,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::BlockMetaInfo;
 use common_expression::BlockMetaInfoPtr;
-use common_storages_pruner::BlockMetaIndex;
+use storages_common_pruner::BlockMetaIndex;
 use storages_common_table_meta::meta::BlockMeta;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
