+++ src/query/storages/fuse/src/pruning/pruner.rs
@@ -22,8 +22,8 @@ use common_expression::Domain;
 use common_expression::Expr;
 use common_expression::TableSchemaRef;
 use common_functions::scalars::BUILTIN_FUNCTIONS;
-use common_storages_index::BlockFilter;
 use opendal::Operator;
+use storages_common_index::BlockFilter;
 use storages_common_table_meta::meta::Location;
 
 use crate::io::BlockFilterReader;
@@ -170,7 +170,7 @@ pub fn new_filter_pruner(
 
 mod util {
     use common_exception::ErrorCode;
-    use common_storages_index::FilterEvalResult;
+    use storages_common_index::FilterEvalResult;
 
     use super::*;
     #[tracing::instrument(level = "debug", skip_all)]
