+++ src/query/storages/fuse/src/pruning/pruning_executor.rs
@@ -27,14 +27,14 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::TableSchemaRef;
 use common_functions::scalars::BUILTIN_FUNCTIONS;
-use common_storages_pruner::BlockMetaIndex;
-use common_storages_pruner::LimiterPruner;
-use common_storages_pruner::LimiterPrunerCreator;
-use common_storages_pruner::RangePruner;
-use common_storages_pruner::RangePrunerCreator;
-use common_storages_pruner::TopNPrunner;
 use futures::future;
 use opendal::Operator;
+use storages_common_pruner::BlockMetaIndex;
+use storages_common_pruner::LimiterPruner;
+use storages_common_pruner::LimiterPrunerCreator;
+use storages_common_pruner::RangePruner;
+use storages_common_pruner::RangePrunerCreator;
+use storages_common_pruner::TopNPrunner;
 use storages_common_table_meta::caches::LoadParams;
 use storages_common_table_meta::meta::BlockMeta;
 use storages_common_table_meta::meta::Location;
