+++ src/query/storages/hive/hive/Cargo.toml
@@ -24,8 +24,8 @@ common-meta-types = { path = "../../../../meta/types" }
 common-pipeline-core = { path = "../../../pipeline/core" }
 common-pipeline-sources = { path = "../../../pipeline/sources" }
 common-storage = { path = "../../../../common/storage" }
-common-storages-index = { path = "../../index" }
 
+storages-common-index = { path = "../../common/index" }
 storages-common-table-meta = { path = "../../common/table-meta" }
 
 async-recursion = "1.0.0"
